Working code for yaw and full control

+++ in4073.h
@@ -36,7 +36,7 @@ uint8_t PreviousMode;
 uint32_t CalibrationStartTime;
 #define CALIBRATION_TIME_US 3000000
 int8_t LRPY[4];
-int8_t k_LRPY[7];
+int16_t k_LRPY[7];
 int16_t LRPY16[4];
 
 bool demo_done;
